Silence warnings

+++ RichTextEditor/Source/RichTextEditor.m
@@ -83,7 +83,6 @@
 	self.toolBar = [[RichTextEditorToolbar alloc] initWithFrame:CGRectMake(0, 0, [self currentScreenBoundsDependOnOrientation].size.width, RICHTEXTEDITOR_TOOLBAR_HEIGHT)
 													   delegate:self
 													 dataSource:self];
-	
 	self.typingAttributesInProgress = NO;
 	self.defaultIndentationSize = 15;
 	
@@ -379,7 +378,7 @@
 	}
 	else
 	{
-		int location = [self offsetFromPosition:self.beginningOfDocument toPosition:self.selectedTextRange.start];
+		NSInteger location = [self offsetFromPosition:self.beginningOfDocument toPosition:self.selectedTextRange.start];
 		
 		if (location == self.text.length)
 			location --;
